Added a new action (USE_KEY) to the action class.
Updated GameManager class to be able to do the following
-keep track of defeated enemies
-check weither the player can pick up item
-keep track of last entered or previos location
-keep track of the gem order for getting the key.

Modified the Location class to be able to get an item in a location.

Updated player class to enable player to monitor quests.

Updated the MainWindow.fxml class to incorproate the following
-Monitor and handle item click for dropping items
-Properly bind all game items to their respective controlls
-Load item images.
-Check for player Win or Loose.

Issue [Game does not start due to bug.]
      [Bug could be because files are not been appropriately read to the game/by the fileReader]

Added ther following to the ViewModel
-booleanProperties for all game items and their respective getters.
-Also improved method funtionality and implementation
-Removed setSelectedAction